requirements.txt
PyYAML~=6.0
tornado~=6.3.3
setuptools~=68.2.0
viswaternet~=1.1.0
matplotlib~=3.5.0